requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
MouseInfo==0.1.3
pillow==10.2.0
PyAutoGUI==0.9.54
pydantic==2.6.4
pydantic_core==2.16.3
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
redis==5.0.3
requests==2.31.0
six==1.16.0
sniffio==1.3.1
spotipy==2.23.0
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
